requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpie==2.4.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Pygments==2.9.0
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.1
urllib3==1.26.4
